fixed print.summary Pr(>|z|) vs. Pr(>|t|) problem
summary.speedglm produces an object whose last column is always a p-value, but sometimes it is named Pr(>|t|) and sometimes it is named Pr(>|z|).  This causes an error with cbind, causing the print to fail.  I fixed it by telling it to use the last column of x$coefficients, x$coefficients[, ncol(x$coefficients)].